Move StableSelectingNetwork out of lib
